+++ NEWS.md
@@ -1,3 +1,8 @@
+# staccuracy 0.2.0
+
+* Added {sa_diff()} function for bootstrapped-based comparison of staccuracies.
+* var_type() is no longer exported since it is really an internal function.
+
 # staccuracy 0.1.0
 
 * Initial CRAN submission.